Move User Settings Tests into their own file
